+++ src/shard_group/src/storage/mod.rs
@@ -1,0 +1,7 @@
+mod group_storage;
+mod index;
+mod state;
+
+pub use group_storage::*;
+pub use index::*;
+pub use state::*;